+++ bin/ntx-builder/src/db/migrations.rs
@@ -30,7 +30,7 @@ mod tests {
     use super::*;
 
     const EXPECTED_SCHEMA_HASHES: [SchemaHash; 1] = [SchemaHash::from_hex(
-        "892f3fb597808a97bdb55762a6ebd4b7941c855d22eb5e0d9b210901720e1125",
+        "e7383731af6f594a2f84ea8c3863325f0219899cff13e1396630c4ea8fed8157",
     )];
 
     #[test]